Move BlockIdSequence to its own file
stack-info: PR: https://github.com/pytorch-labs/helion/pull/129, branch: jansel/stack/25
